requirements.txt
six
Faker
pillow
markdown
python-dateutil
pytz
DateTime

django==4.2
django-environ
django-filter
django-allauth==0.55.0
django-resized
python-decouple
django-ckeditor
django-crispy-forms
djangorestframework
dj-rest-auth

crispy-bootstrap5
django-browser-reload
drf-yasg
django-phonenumber-field[phonenumbers]